# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API untuk sistem manajemen Wira Teknik. Mendukung autentikasi, manajemen pengguna, dan fitur bisnis lainnya.",
        "title": "Wira Teknik API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Tim Wira Teknik",
            "email": "admin@wira-teknik.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar semua pesanan beserta item-itemnya",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Ambil semua pesanan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Order"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat pesanan baru dengan item. PPN 11% dan subtotal dihitung otomatis. remaining_qty diset sama dengan order_qty.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Buat pesanan baru",
                "parameters": [
                    {
                        "description": "Data pesanan",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Order"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail pesanan termasuk items dan shipments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Ambil detail pesanan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Order"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengupdate data header pesanan (hanya jika status masih pending)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Update header pesanan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data pesanan",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Order"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus pesanan beserta item-itemnya (hanya jika status masih pending)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Hapus pesanan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.OrderItemRequestPayload"
                    }
                },
                "order_date": {
                    "type": "string",
                    "example": "2026-05-09"
                },
                "po_no": {
                    "type": "string",
                    "example": "PO-2026-001"
                },
                "recipient_address": {
                    "type": "string",
                    "example": "Jl. Industri No. 10"
                },
                "recipient_email": {
                    "type": "string",
                    "example": "purchasing@majujaya.com"
                },
                "recipient_name": {
                    "type": "string",
                    "example": "PT Maju Jaya"
                },
                "recipient_phone": {
                    "type": "string",
                    "example": "081234567890"
                }
            }
        },
...
        "controllers.OrderItemRequestPayload": {
            "type": "object",
            "properties": {
                "order_qty": {
                    "type": "integer",
                    "example": 100
                },
                "product_name": {
                    "type": "string",
                    "example": "Bearing SKF 6205"
                },
                "unit_price": {
                    "type": "number",
                    "example": 75000
                }
            }
        },
...
        "controllers.UpdateOrderRequest": {
            "type": "object",
            "properties": {
                "po_no": {
                    "type": "string",
                    "example": "PO-2026-001"
                },
                "recipient_address": {
                    "type": "string",
                    "example": "Jl. Industri No. 10"
                },
                "recipient_email": {
                    "type": "string",
                    "example": "purchasing@majujaya.com"
                },
                "recipient_name": {
                    "type": "string",
                    "example": "PT Maju Jaya"
                },
                "recipient_phone": {
                    "type": "string",
                    "example": "081234567890"
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderItem"
                    }
                },
                "order_date": {
                    "type": "string"
                },
                "order_status": {
                    "$ref": "#/definitions/models.OrderStatus"
                },
                "po_no": {
                    "type": "string"
                },
                "recipient_address": {
                    "type": "string"
                },
                "recipient_email": {
                    "type": "string"
                },
                "recipient_name": {
                    "type": "string"
                },
                "recipient_phone": {
                    "type": "string"
                },
                "shipments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Shipment"
                    }
                },
                "transaction_no": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.OrderItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "order_qty": {
                    "type": "integer"
                },
                "ppn": {
                    "type": "number"
                },
                "product_name": {
                    "type": "string"
                },
                "remaining_qty": {
                    "type": "integer"
                },
                "subtotal": {
                    "type": "number"
                },
                "unit_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.OrderStatus": {
            "type": "string",
            "enum": [
                "pending",
                "partial",
                "shipped",
                "completed"
            ],
            "x-enum-varnames": [
                "OrderStatusPending",
                "OrderStatusPartial",
                "OrderStatusShipped",
                "OrderStatusCompleted"
            ]
        },
...
        "models.Shipment": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ShipmentItem"
                    }
                },
                "order_id": {
                    "type": "string"
                },
                "received_date": {
                    "type": "string"
                },
                "shipping_date": {
                    "type": "string"
                },
                "shipping_status": {
                    "$ref": "#/definitions/models.ShippingStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ShipmentItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order_item_id": {
                    "type": "string"
                },
                "shipment_id": {
                    "type": "string"
                },
                "shipping_qty": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ShippingStatus": {
            "type": "string",
            "enum": [
                "dikirim",
                "diterima"
            ],
            "x-enum-varnames": [
                "ShippingStatusDikirim",
                "ShippingStatusDiterima"
            ]
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operasi berhasil"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}